+++ decomp/ST.exe/functions/00604350/decomp.c
@@ -83,35 +83,34 @@ int __fastcall FUN_00604350(AnonShape_00604350_448AE2F1 *param_1)
             ((thunk_FUN_00558c00(g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,iVar1,
                                  iVar5,&local_18,&local_1c), pVVar4 = g_visibleClass_00802A88,
              local_8 < 0 ||
-             ((((4 < local_8 || (local_18 < 0)) || ((int)pVVar3->field_0030 <= local_18)) ||
+             ((((4 < local_8 || (local_18 < 0)) || (pVVar3->field_0030 <= local_18)) ||
               ((iVar5 = (&DAT_0079aed0)[local_8] + local_1c, iVar5 < 0 ||
-               ((int)pVVar3->field_0034 <= iVar5)))))))) || (pVVar3->field_004C == 0)) ||
+               (pVVar3->field_0034 <= iVar5)))))))) || (pVVar3->field_004C == 0)) ||
           ((*(char *)(iVar5 * pVVar3->field_0030 + pVVar3->field_004C + local_18) != '\0' ||
            (DAT_0080874d == -1)))) ||
          (((((g_visibleClass_00802A88->field_00F8 == 0 ||
              (((thunk_FUN_00558c00(g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,
                                    local_c,iVar7,&local_1c,&local_18),
-               pVVar3 = g_visibleClass_00802A88, local_1c < 0 ||
-               ((int)pVVar4->field_0030 <= local_1c)) ||
-              (iVar5 = (&DAT_0079aed0)[local_8] + local_18, iVar5 < 0)))) ||
-            ((((int)pVVar4->field_0034 <= iVar5 || (pVVar4->field_004C == 0)) ||
+               pVVar3 = g_visibleClass_00802A88, local_1c < 0 || (pVVar4->field_0030 <= local_1c))
+              || (iVar5 = (&DAT_0079aed0)[local_8] + local_18, iVar5 < 0)))) ||
+            (((pVVar4->field_0034 <= iVar5 || (pVVar4->field_004C == 0)) ||
              (*(char *)(iVar5 * pVVar4->field_0030 + pVVar4->field_004C + local_1c) != '\0')))) ||
            ((((DAT_0080874d == -1 || (g_visibleClass_00802A88->field_00F8 == 0)) ||
              ((thunk_FUN_00558c00(g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,
                                   local_10,iVar7,&local_1c,&local_18), iVar1 = local_8,
               iVar5 = local_14, pVVar4 = g_visibleClass_00802A88, local_1c < 0 ||
-              ((((int)pVVar3->field_0030 <= local_1c ||
+              (((pVVar3->field_0030 <= local_1c ||
                 (iVar7 = (&DAT_0079aed0)[local_8] + local_18, iVar7 < 0)) ||
-               ((int)pVVar3->field_0034 <= iVar7)))))) ||
+               (pVVar3->field_0034 <= iVar7)))))) ||
             ((pVVar3->field_004C == 0 ||
              (*(char *)(iVar7 * pVVar3->field_0030 + pVVar3->field_004C + local_1c) != '\0')))))) ||
           ((DAT_0080874d == -1 ||
            (((g_visibleClass_00802A88->field_00F8 == 0 ||
              (thunk_FUN_00558c00(g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,local_c
                                  ,local_14,&local_1c,&local_18), local_1c < 0)) ||
-            (((int)pVVar4->field_0030 <= local_1c ||
+            ((pVVar4->field_0030 <= local_1c ||
              ((((local_18 = (&DAT_0079aed0)[iVar1] + local_18, local_18 < 0 ||
-                ((int)pVVar4->field_0034 <= local_18)) || (pVVar4->field_004C == 0)) ||
+                (pVVar4->field_0034 <= local_18)) || (pVVar4->field_004C == 0)) ||
               ((*(char *)(local_18 * pVVar4->field_0030 + pVVar4->field_004C + local_1c) != '\0' ||
                (iVar5 = thunk_FUN_005f1d80(g_visibleClass_00802A88,local_10,iVar5,local_8),
                iVar5 == 1)))))))))))))) {
